add java dependancy needed for exporting to html
When exporting the game to HTML5, java is needed due to some js files it
produces. Because of that, there was a switch to endlessm.Sdk as the
runtime, as it had most of the tools we were needing anyways. killall
was not needed after that, as its shipped with endlessm.Sdk.